+++ pandas/tests/frame/test_to_csv.py
@@ -893,14 +893,16 @@ class TestDataFrameToCSV(TestData):
 
     def test_to_csv_from_csv_categorical(self):
 
-        # CSV with categoricals should result in the same output as when one
-        # would add a "normal" Series/DataFrame.
+        # CSV with Categoricals should result in the same output
+        # as when one would add a "normal" Series/DataFrame.
         s = Series(pd.Categorical(['a', 'b', 'b', 'a', 'a', 'c', 'c', 'c']))
         s2 = Series(['a', 'b', 'b', 'a', 'a', 'c', 'c', 'c'])
         res = StringIO()
-        s.to_csv(res)
+
+        s.to_csv(res, header=False)
         exp = StringIO()
-        s2.to_csv(exp)
+
+        s2.to_csv(exp, header=False)
         assert res.getvalue() == exp.getvalue()
 
         df = DataFrame({"s": s})
